Split global layout to index.py
