+++ webmc3/univariate/__init__.py
@@ -1,1 +1,2 @@
+from .components import add_callbacks
 from .layout import layout
